Break apart the meeting link component monstrosity.
