+++ packages/webapp/src/elections/components/ongoing-election-components/meeting-link/index.ts
@@ -1,0 +1,1 @@
+export * from "./meeting-link";
