+++ packages/webapp/src/elections/components/ongoing-election-components/meeting-link/meeting-link.md
@@ -1,4 +1,4 @@
-# `<RequestElectionMeetingLinkButton />`
+# `<MeetingLink />`
 
 ## Notes
