+++ packages/webapp/src/elections/components/ongoing-election-components/ongoing-round-segment.tsx
@@ -30,7 +30,7 @@ import {
 
 import Consensometer from "./consensometer";
 import RoundHeader from "./round-header";
-import { RequestElectionMeetingLinkButton } from "./request-election-meeting-link-button";
+import MeetingLink from "./meeting-link/meeting-link";
 import VotingRoundParticipants from "./voting-round-participants";
 import { setVote } from "../../transactions";
 
@@ -227,7 +227,7 @@ export const OngoingRoundSegment = ({
                 {[RoundStage.PreMeeting, RoundStage.Meeting].includes(
                     stage
                 ) && (
-                    <RequestElectionMeetingLinkButton
+                    <MeetingLink
                         stage={stage}
                         roundIndex={roundIndex}
                         meetingStartTime={meetingStartTime}
